Remove need to guard serverinterface and repeat file loading logic
